+++ src/com/tigerknows/ui/hotel/HotelOrderListFragment.java
@@ -574,7 +574,7 @@ public class HotelOrderListFragment extends BaseFragment implements View.OnClick
         
         final HotelOrderOperation hotelOrderOperation = (HotelOrderOperation)(tkAsyncTask.getBaseQuery());
         HotelOrderStatesResponse response = (HotelOrderStatesResponse) hotelOrderOperation.getResponse();
-        List<Integer> states = response.getStates();
+        List<Long> states = response.getStates();
         
         LogWrapper.i(TAG, "Number of states get: " + states.size());
         
@@ -596,7 +596,7 @@ public class HotelOrderListFragment extends BaseFragment implements View.OnClick
      * @param ids
      * @param orders
      */
-    public void updateOrderState(List<Integer> states, String ids, List<HotelOrder> orders){
+    public void updateOrderState(List<Long> states, String ids, List<HotelOrder> orders){
     	
     	updateOrderInMemory(states, ids, orders);
     	
@@ -610,7 +610,7 @@ public class HotelOrderListFragment extends BaseFragment implements View.OnClick
      * @param ids
      * @param orders
      */
-    private void updateOrderInMemory(List<Integer> states, String ids, List<HotelOrder> orders){
+    private void updateOrderInMemory(List<Long> states, String ids, List<HotelOrder> orders){
 
         long stateUpdateTime = System.currentTimeMillis();
     	String[] idArray = ids.split("_");
@@ -628,7 +628,7 @@ public class HotelOrderListFragment extends BaseFragment implements View.OnClick
     				for (int j = idArray.length-1; j >= 0; j--) {
     					if(idArray[j].equals( order.getId()) ){
     						if(states.get(j)!=HotelOrder.STATE_NONE && j<states.size()){
-    							order.setState(states.get(j));
+    							order.setState(states.get(j).intValue());
     							order.setStateUpdateTime(stateUpdateTime);
     							ordersToUpdateToDB.addLast(order);
     							ordersQuerying.remove(order);
